refactor: move application and domain layer to core directory
