+++ apps/backend/src/core/application/dto/near-pet.dto.ts
