+++ apps/backend/src/core/application/repositories/pet.repository.ts
