+++ apps/backend/src/core/application/usecases/owner/update-posted-pet/update-posted-pet.ts
@@ -2,8 +2,8 @@ import { Inject, Injectable } from '@nestjs/common';
 import PetRepository, {
   PET_REPOSITORY,
 } from '../../../repositories/pet.repository';
-import { UserService } from '../../../../user/user.service';
-import { UserRepository } from '../../../../user/repositories/user-repository.interface';
+import { UserService } from '../../../../../user/user.service';
+import { UserRepository } from '../../../../../user/repositories/user-repository.interface';
 import { UnauthorizedError } from '../../../../domain/errors/unauthorized.error';
 import { NotFoundError } from '../../../../domain/errors/not-found.error';
 import type { Input as UpdatablePetData } from '../post-pet-for-adoption/post-pet-for-adoption';
