+++ apps/backend/src/core/domain/errors/entity.error.ts
