+++ apps/backend/src/core/domain/errors/not-found.error.ts
