+++ apps/backend/src/infra/controllers/geolocation.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, Get, Query } from '@nestjs/common';
-import { GeolocationService } from '../../application/dao/geolocation.service.abstract';
+import { GeolocationService } from '../../core/application/dao/geolocation.service.abstract';
 
 @Controller('api/v1/location')
 export class GeolocationController {
