+++ apps/backend/src/modules/pet.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
 import { PetsController } from '../infra/controllers/pets.controller';
 import { UserModule } from './user.module';
-import PostPetForAdoption from '../application/usecases/owner/post-pet-for-adoption/post-pet-for-adoption';
-import FindNearestPets from '../application/usecases/adopter/find-nearest-pets/find-nearest-pets';
-import UpdatePostedPet from '../application/usecases/owner/update-posted-pet/update-posted-pet';
-import RemovePostedPet from '../application/usecases/owner/remove-posted-pet/remove-posted-pet';
+import PostPetForAdoption from '../core/application/usecases/owner/post-pet-for-adoption/post-pet-for-adoption';
+import FindNearestPets from '../core/application/usecases/adopter/find-nearest-pets/find-nearest-pets';
+import UpdatePostedPet from '../core/application/usecases/owner/update-posted-pet/update-posted-pet';
+import RemovePostedPet from '../core/application/usecases/owner/remove-posted-pet/remove-posted-pet';
 
 @Module({
   imports: [UserModule],
